requirements.txt
streamlit
pymupdf
python-pptx
python-docx
matplotlib
scikit-learn
reportlab
openai